refactor(#17): reorganize src/ to Waaseyaa Application Directory Convention v1.0
Move files to convention directories (Domain/, Support/, Provider/,
Ingestion/Pipeline/) and update all namespaces, use statements, and
composer.json provider path. No business logic changes.
