+++ composer.json
@@ -36,7 +36,7 @@
     "extra": {
         "waaseyaa": {
             "providers": [
-                "MyClaudia\\McClaudiaServiceProvider"
+                "MyClaudia\\Provider\\McClaudiaServiceProvider"
             ]
         }
     },
