+++ src/Controller/DayBriefController.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace MyClaudia\Controller;
 
-use MyClaudia\DayBrief\BriefSessionStore;
+use MyClaudia\Domain\DayBrief\Service\BriefSessionStore;
 use Waaseyaa\Entity\EntityTypeManager;
 use Waaseyaa\SSR\SsrResponse;
