+++ tests/Unit/DayBrief/BriefSessionStoreTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace MyClaudia\Tests\Unit\DayBrief;
 
-use MyClaudia\DayBrief\BriefSessionStore;
+use MyClaudia\Domain\DayBrief\Service\BriefSessionStore;
 use PHPUnit\Framework\TestCase;
 
 final class BriefSessionStoreTest extends TestCase
